+++ src/context/simulation_context.cpp
@@ -32,7 +32,7 @@
 #include "symmetry/crystal_symmetry.hpp"
 #include "symmetry/check_gvec.hpp"
 #include "potential/xc_functional.hpp"
-#include "linalg/linalg_spla.hpp"
+#include "core/la/linalg_spla.hpp"
 #include "lapw/step_function.hpp"
 
 namespace sirius {
