+++ src/core/la/dmatrix.cpp
