+++ src/core/la/eigenproblem.cpp
