+++ src/core/la/eigenproblem.hpp
