+++ src/core/la/eigensolver.cpp
