+++ src/core/la/inverse_sqrt.hpp
@@ -1,12 +1,38 @@
+// Copyright (c) 2013-2023 Anton Kozhevnikov, Thomas Schulthess
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without modification, are permitted provided that
+// the following conditions are met:
+//
+// 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the
+//    following disclaimer.
+// 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions
+//    and the following disclaimer in the documentation and/or other materials provided with the distribution.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
+// WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+// ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+// OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+/** \file inverse_sqrt.hpp
+ *
+ *  \brief Compute inverse square root of the matrix.
+ */
+
 #ifndef __INVERSE_SQRT_HPP__
 #define __INVERSE_SQRT_HPP__
 
-#include "linalg/dmatrix.hpp"
-#include "linalg/eigensolver.hpp"
-#include "linalg/linalg.hpp"
+#include "core/la/dmatrix.hpp"
+#include "core/la/eigensolver.hpp"
+#include "core/la/linalg.hpp"
 #include "core/rte/rte.hpp"
 
 namespace sirius {
+
+namespace la {
 
 /// Compute inverse square root of the matrix.
 /** As by-product, return the eigen-vectors and the eigen-values of the matrix. */
@@ -53,4 +79,5 @@ inverse_sqrt(la::dmatrix<T>& A__, int N__)
 
 }
 
+}
 #endif
