+++ src/core/la/linalg.hpp
